% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happinessData.r
\docType{data}
\name{happinessData}
\alias{happinessData}
\title{An analysis of the 2021 World Happiness Report.}
\format{
a data frame with 1667 observations
\describe{
  \item{Country.name}{The name of the country}
  \item{year}{The year that the corresponding data was evaluated}
  \item{Log.GDP.per.capita}{The Gross Domestic Product score of that country
                           adjusted for the population}
  \item{Social.support}{A score quantifying the average person's trust in
       their social environment determined by the Sustainable Development
       Solutions Network (by the United Nations)}
  \item{Healthy.life.expectancy.at.birth}{A score quantifying the average
       person's life expectancy at birth determined by the Sustainable
       Development Solutions Network (by the United Nations)}
  \item{Freedom.to.make.life.choices}{A score quantifying the average
       person's freedom to make life choices determined by the Sustainable
       Development Solutions Network (by the United Nations)}
  \item{Generosity}{A score quantifying the average person's generosity in
       their social environment determined by the Sustainable Development
       Solutions Network (by the United Nations)}
  \item{Percetptions.of.corruption}{A score quantifying the average person's
       trust in their government determined by the Sustainable Development
       Solutions Network (by the United Nations)}
  \item{Regional.indicator}{A broader geographical area that a country is
       located in (ie residing continent)}
}
}
\source{
\url{https://www.kaggle.com/datasets/ajaypalsinghlo/world-happiness-report-2021}
}
\usage{
happinessData
}
\description{
An analysis of the 2021 World Happiness Report.
}
\keyword{datasets}
